test(network): move lib tests to tests dir
