+++ network/src/lib.rs
@@ -252,456 +252,3 @@ impl Server {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::external::message_types::*;
-    use snarkos_testing::{
-        consensus::{BLOCK_1, BLOCK_1_HEADER_HASH, BLOCK_2, BLOCK_2_HEADER_HASH, FIXTURE_VK, TEST_CONSENSUS},
-        dpc::load_verifying_parameters,
-        network::{read_header, read_payload},
-    };
-    use snarkvm_objects::block_header_hash::BlockHeaderHash;
-
-    use std::{sync::Arc, time::Duration};
-
-    use chrono::Utc;
-    use parking_lot::{Mutex, RwLock};
-    use tokio::{
-        io::{AsyncReadExt, AsyncWriteExt},
-        net::{TcpListener, TcpStream},
-    };
-
-    async fn test_node(
-        bootnodes: Vec<String>,
-        peer_sync_interval: Duration,
-        block_sync_interval: Duration,
-        transaction_sync_interval: Duration,
-    ) -> Server {
-        let storage = FIXTURE_VK.ledger();
-        let memory_pool = snarkos_consensus::MemoryPool::new();
-        let memory_pool_lock = Arc::new(Mutex::new(memory_pool));
-        let consensus = TEST_CONSENSUS.clone();
-        let parameters = load_verifying_parameters();
-        let socket_address = None;
-        let min_peers = 1;
-        let max_peers = 10;
-        let is_bootnode = false;
-        let is_miner = false;
-
-        let environment = Environment::new(
-            Arc::new(RwLock::new(storage)),
-            memory_pool_lock,
-            Arc::new(consensus),
-            Arc::new(parameters),
-            socket_address,
-            min_peers,
-            max_peers,
-            bootnodes,
-            is_bootnode,
-            is_miner,
-            peer_sync_interval,
-            block_sync_interval,
-            transaction_sync_interval,
-        )
-        .unwrap();
-
-        Server::new(environment).await.unwrap()
-    }
-
-    async fn write_message_to_stream(payload: Payload, peer_stream: &mut TcpStream) {
-        let payload = bincode::serialize(&payload).unwrap();
-        let header = MessageHeader {
-            len: payload.len() as u32,
-        }
-        .as_bytes();
-        peer_stream.write_all(&header[..]).await.unwrap();
-        peer_stream.write_all(&payload).await.unwrap();
-        peer_stream.flush().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn starts_server() {
-        let mut server = test_node(vec![], Duration::new(10, 0), Duration::new(10, 0), Duration::new(10, 0)).await;
-        assert!(server.start().await.is_ok());
-        let address = server.local_address().unwrap();
-
-        assert!(TcpListener::bind(address).await.is_err());
-        assert_eq!(server.peers.number_of_connected_peers(), 0);
-    }
-
-    #[tokio::test]
-    async fn handshake_responder_side() {
-        // start a test node and listen for incoming connections
-        let mut node = test_node(vec![], Duration::new(10, 0), Duration::new(10, 0), Duration::new(10, 0)).await;
-        node.start().await.unwrap();
-        let node_listener = node.local_address().unwrap();
-
-        // set up a fake node (peer), which is just a socket
-        let mut peer_stream = TcpStream::connect(&node_listener).await.unwrap();
-
-        // register the addresses bound to the connection between the node and the peer
-        let peer_address = peer_stream.local_addr().unwrap();
-
-        // the peer initiates a handshake by sending a Version message
-        let version = Payload::Version(Version::new(1u64, 1u32, 1u64, peer_address.port()));
-        write_message_to_stream(version, &mut peer_stream).await;
-
-        // at this point the node should have marked the peer as ' connecting'
-        sleep(Duration::from_millis(200)).await;
-        assert!(node.peers.is_connecting(&peer_address));
-
-        // the buffer for peer's reads
-        let mut peer_buf = [0u8; 64];
-
-        // check if the peer has received the Verack message from the node
-        let len = read_header(&mut peer_stream).await.unwrap().len();
-        let payload = read_payload(&mut peer_stream, &mut peer_buf[..len]).await.unwrap();
-        assert!(matches!(bincode::deserialize(&payload).unwrap(), Payload::Verack(..)));
-
-        // check if it was followed by a Version message
-        let len = read_header(&mut peer_stream).await.unwrap().len();
-        let payload = read_payload(&mut peer_stream, &mut peer_buf[..len]).await.unwrap();
-        let version = if let Payload::Version(version) = bincode::deserialize(&payload).unwrap() {
-            version
-        } else {
-            unreachable!();
-        };
-
-        // in response to the Version, the peer sends a Verack message to finish the handshake
-        let verack = Payload::Verack(Verack::new(version.nonce));
-        write_message_to_stream(verack, &mut peer_stream).await;
-
-        // the node should now have register the peer as 'connected'
-        sleep(Duration::from_millis(200)).await;
-        assert!(node.peers.is_connected(&peer_address));
-        assert_eq!(node.peers.number_of_connected_peers(), 1);
-    }
-
-    #[tokio::test]
-    async fn handshake_initiator_side() {
-        // start a fake peer which is just a socket
-        let peer_listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let peer_address = peer_listener.local_addr().unwrap();
-
-        // start node with the peer as a bootnode; that way it will get connected to
-        let mut node = test_node(
-            vec![peer_address.to_string()],
-            Duration::new(10, 0),
-            Duration::new(10, 0),
-            Duration::new(10, 0),
-        )
-        .await;
-        node.start().await.unwrap();
-
-        // accept the node's connection on peer side
-        let (mut peer_stream, _node_address) = peer_listener.accept().await.unwrap();
-
-        // the buffer for peer's reads
-        let mut peer_buf = [0u8; 64];
-
-        // the peer should receive a Version message from the node (initiator of the handshake)
-        let len = read_header(&mut peer_stream).await.unwrap().len();
-        let payload = read_payload(&mut peer_stream, &mut peer_buf[..len]).await.unwrap();
-        let version = if let Payload::Version(version) = bincode::deserialize(&payload).unwrap() {
-            version
-        } else {
-            unreachable!();
-        };
-
-        // at this point the node should have marked the peer as 'connecting'
-        assert!(node.peers.is_connecting(&peer_address));
-
-        // the peer responds with a Verack acknowledging the Version message
-        let verack = Payload::Verack(Verack::new(version.nonce));
-        write_message_to_stream(verack, &mut peer_stream).await;
-
-        // the peer then follows up with a Version message
-        let version = Payload::Version(Version::new(1u64, 1u32, 1u64, peer_address.port()));
-        write_message_to_stream(version, &mut peer_stream).await;
-
-        // the node should now have registered the peer as 'connected'
-        sleep(Duration::from_millis(200)).await;
-        assert!(node.peers.is_connected(&peer_address));
-        assert_eq!(node.peers.number_of_connected_peers(), 1);
-    }
-
-    async fn assert_node_rejected_message(node: &Server, peer_stream: &mut TcpStream) {
-        // slight delay for server to process the message
-        sleep(Duration::from_millis(200)).await;
-
-        // read the response from the stream
-        let mut buffer = String::new();
-        let bytes_read = peer_stream.read_to_string(&mut buffer).await.unwrap();
-
-        // check node's response is empty
-        assert_eq!(bytes_read, 0);
-        assert!(buffer.is_empty());
-
-        // check the node's state hasn't been altered by the message
-        assert!(!node.peers.is_connecting(&peer_stream.local_addr().unwrap()));
-        assert_eq!(node.peers.number_of_connected_peers(), 0);
-    }
-
-    #[tokio::test]
-    async fn reject_non_version_messages_before_handshake() {
-        // start the node
-        let mut node = test_node(vec![], Duration::new(10, 0), Duration::new(10, 0), Duration::new(10, 0)).await;
-        node.start().await.unwrap();
-
-        // start the fake node (peer) which is just a socket
-        // note: the connection needs to be re-established as it is reset
-        let mut peer_stream = TcpStream::connect(node.local_address().unwrap()).await.unwrap();
-
-        // send a GetPeers message without a prior handshake established
-        write_message_to_stream(Payload::GetPeers, &mut peer_stream).await;
-
-        // verify the node rejected the message, the response to the peer is empty and the node's
-        // state is unaltered
-        assert_node_rejected_message(&node, &mut peer_stream).await;
-
-        // GetMemoryPool
-        let mut peer_stream = TcpStream::connect(node.local_address().unwrap()).await.unwrap();
-        write_message_to_stream(Payload::GetMemoryPool, &mut peer_stream).await;
-        assert_node_rejected_message(&node, &mut peer_stream).await;
-
-        // GetBlock
-        let mut peer_stream = TcpStream::connect(node.local_address().unwrap()).await.unwrap();
-        let block_hash = BlockHeaderHash::new([0u8; 32].to_vec());
-        write_message_to_stream(Payload::GetBlock(block_hash), &mut peer_stream).await;
-        assert_node_rejected_message(&node, &mut peer_stream).await;
-
-        // GetSync
-        let mut peer_stream = TcpStream::connect(node.local_address().unwrap()).await.unwrap();
-        let block_hash = BlockHeaderHash::new([0u8; 32].to_vec());
-        write_message_to_stream(Payload::GetSync(vec![block_hash]), &mut peer_stream).await;
-        assert_node_rejected_message(&node, &mut peer_stream).await;
-
-        // Peers
-        let mut peer_stream = TcpStream::connect(node.local_address().unwrap()).await.unwrap();
-        let peers = vec![("127.0.0.1:0".parse().unwrap(), Utc::now())];
-        write_message_to_stream(Payload::Peers(peers), &mut peer_stream).await;
-        assert_node_rejected_message(&node, &mut peer_stream).await;
-
-        // MemoryPool
-        let mut peer_stream = TcpStream::connect(node.local_address().unwrap()).await.unwrap();
-        let memory_pool = vec![vec![0u8, 10]];
-        write_message_to_stream(Payload::MemoryPool(memory_pool), &mut peer_stream).await;
-        assert_node_rejected_message(&node, &mut peer_stream).await;
-
-        // Block
-        let mut peer_stream = TcpStream::connect(node.local_address().unwrap()).await.unwrap();
-        let block = vec![0u8, 10];
-        write_message_to_stream(Payload::Block(block), &mut peer_stream).await;
-        assert_node_rejected_message(&node, &mut peer_stream).await;
-
-        // SyncBlock
-        let mut peer_stream = TcpStream::connect(node.local_address().unwrap()).await.unwrap();
-        let sync_block = vec![0u8, 10];
-        write_message_to_stream(Payload::SyncBlock(sync_block), &mut peer_stream).await;
-        assert_node_rejected_message(&node, &mut peer_stream).await;
-
-        // Sync
-        let mut peer_stream = TcpStream::connect(node.local_address().unwrap()).await.unwrap();
-        let block_hash = BlockHeaderHash::new(vec![0u8; 32]);
-        write_message_to_stream(Payload::Sync(vec![block_hash]), &mut peer_stream).await;
-        assert_node_rejected_message(&node, &mut peer_stream).await;
-
-        // Transaction
-        let mut peer_stream = TcpStream::connect(node.local_address().unwrap()).await.unwrap();
-        let transaction = vec![0u8, 10];
-        write_message_to_stream(Payload::Transaction(transaction), &mut peer_stream).await;
-        assert_node_rejected_message(&node, &mut peer_stream).await;
-
-        // Verack
-        let mut peer_stream = TcpStream::connect(node.local_address().unwrap()).await.unwrap();
-        let verack = Verack::new(1u64);
-        write_message_to_stream(Payload::Verack(verack), &mut peer_stream).await;
-        assert_node_rejected_message(&node, &mut peer_stream).await;
-    }
-
-    async fn handshake(
-        peer_sync_interval: Duration,
-        block_sync_interval: Duration,
-        transaction_sync_interval: Duration,
-    ) -> (Server, TcpStream) {
-        // start a test node and listen for incoming connections
-        let mut node = test_node(
-            vec![],
-            peer_sync_interval,
-            block_sync_interval,
-            transaction_sync_interval,
-        )
-        .await;
-        node.start().await.unwrap();
-        let node_listener = node.local_address().unwrap();
-
-        // set up a fake node (peer), which is just a socket
-        let mut peer_stream = TcpStream::connect(&node_listener).await.unwrap();
-
-        // register the addresses bound to the connection between the node and the peer
-        let peer_address = peer_stream.local_addr().unwrap();
-
-        // the peer initiates a handshake by sending a Version message
-        let version = Payload::Version(Version::new(1u64, 1u32, 1u64, peer_address.port()));
-        write_message_to_stream(version, &mut peer_stream).await;
-
-        // at this point the node should have marked the peer as ' connecting'
-        sleep(Duration::from_millis(200)).await;
-        assert!(node.peers.is_connecting(&peer_address));
-
-        // the buffer for peer's reads
-        let mut peer_buf = [0u8; 64];
-
-        // check if the peer has received the Verack message from the node
-        let len = read_header(&mut peer_stream).await.unwrap().len();
-        let payload = read_payload(&mut peer_stream, &mut peer_buf[..len]).await.unwrap();
-        assert!(matches!(bincode::deserialize(&payload).unwrap(), Payload::Verack(_)));
-
-        // check if it was followed by a Version message
-        let len = read_header(&mut peer_stream).await.unwrap().len();
-        let payload = read_payload(&mut peer_stream, &mut peer_buf[..len]).await.unwrap();
-        let version = if let Payload::Version(version) = bincode::deserialize(&payload).unwrap() {
-            version
-        } else {
-            unreachable!();
-        };
-
-        // in response to the Version, the peer sends a Verack message to finish the handshake
-        let verack = Payload::Verack(Verack::new(version.nonce));
-        write_message_to_stream(verack, &mut peer_stream).await;
-
-        // the node should now have register the peer as 'connected'
-        sleep(Duration::from_millis(200)).await;
-        assert!(node.peers.is_connected(&peer_address));
-
-        (node, peer_stream)
-    }
-
-    #[tokio::test]
-    async fn sync_initiator_side() {
-        // handshake between the fake node and full node
-        let (node, mut peer_stream) =
-            handshake(Duration::new(100, 0), Duration::new(10, 0), Duration::new(100, 0)).await;
-
-        // the buffer for peer's reads
-        let mut peer_buf = [0u8; 64];
-
-        // check GetSync message was received
-        let len = read_header(&mut peer_stream).await.unwrap().len();
-        let payload = read_payload(&mut peer_stream, &mut peer_buf[..len]).await.unwrap();
-        assert!(matches!(bincode::deserialize(&payload).unwrap(), Payload::GetSync(..)));
-
-        let block_1_header_hash = BlockHeaderHash::new(BLOCK_1_HEADER_HASH.to_vec());
-        let block_2_header_hash = BlockHeaderHash::new(BLOCK_2_HEADER_HASH.to_vec());
-
-        let block_header_hashes = vec![block_1_header_hash.clone(), block_2_header_hash.clone()];
-
-        // respond to GetSync with Sync message containing the block header hashes of the missing
-        // blocks
-        let sync = Payload::Sync(block_header_hashes);
-        write_message_to_stream(sync, &mut peer_stream).await;
-
-        // make sure both GetBlock messages are received
-        let len = read_header(&mut peer_stream).await.unwrap().len();
-        let payload = read_payload(&mut peer_stream, &mut peer_buf[..len]).await.unwrap();
-        let block_hash = if let Payload::GetBlock(block_hash) = bincode::deserialize(&payload).unwrap() {
-            block_hash
-        } else {
-            unreachable!();
-        };
-
-        assert_eq!(block_hash, block_1_header_hash);
-
-        let len = read_header(&mut peer_stream).await.unwrap().len();
-        let payload = read_payload(&mut peer_stream, &mut peer_buf[..len]).await.unwrap();
-        let block_hash = if let Payload::GetBlock(block_hash) = bincode::deserialize(&payload).unwrap() {
-            block_hash
-        } else {
-            unreachable!();
-        };
-
-        assert_eq!(block_hash, block_2_header_hash);
-
-        // respond with the full blocks
-        let block_1 = Payload::Block(BLOCK_1.to_vec());
-        write_message_to_stream(block_1, &mut peer_stream).await;
-
-        let block_2 = Payload::Block(BLOCK_2.to_vec());
-        write_message_to_stream(block_2, &mut peer_stream).await;
-
-        sleep(Duration::from_millis(200)).await;
-
-        // check the blocks have been added to the node's chain
-        assert!(
-            node.environment
-                .storage()
-                .read()
-                .block_hash_exists(&block_1_header_hash)
-        );
-
-        assert!(
-            node.environment
-                .storage()
-                .read()
-                .block_hash_exists(&block_2_header_hash)
-        );
-    }
-
-    #[tokio::test]
-    async fn sync_responder_side() {
-        // handshake between the fake and full node
-        let (node, mut peer_stream) =
-            handshake(Duration::new(100, 0), Duration::new(10, 0), Duration::new(100, 0)).await;
-
-        // insert block into node
-        let block_struct_1 = snarkvm_objects::Block::deserialize(&BLOCK_1).unwrap();
-        node.environment
-            .consensus_parameters()
-            .receive_block(
-                node.environment.dpc_parameters(),
-                &node.environment.storage().read(),
-                &mut node.environment.memory_pool().lock(),
-                &block_struct_1,
-            )
-            .unwrap();
-
-        // send a GetSync with an empty vec as only the genesis block is in the ledger
-        let get_sync = Payload::GetSync(vec![]);
-        write_message_to_stream(get_sync, &mut peer_stream).await;
-
-        // the buffer for peer's reads
-        let mut peer_buf = [0u8; 4096];
-
-        // receive a Sync message from the node with the block header
-        let len = read_header(&mut peer_stream).await.unwrap().len();
-        let payload = read_payload(&mut peer_stream, &mut peer_buf[..len]).await.unwrap();
-        let sync = if let Payload::Sync(sync) = bincode::deserialize(&payload).unwrap() {
-            sync
-        } else {
-            unreachable!();
-        };
-
-        let block_header_hash = sync.first().unwrap();
-
-        // check it matches the block inserted into the node's ledger
-        assert_eq!(*block_header_hash, block_struct_1.header.get_hash());
-
-        // request the block from the node
-        let get_block = Payload::GetBlock(block_header_hash.clone());
-        write_message_to_stream(get_block, &mut peer_stream).await;
-
-        // receive a SyncBlock message with the requested block
-        let len = read_header(&mut peer_stream).await.unwrap().len();
-        let payload = read_payload(&mut peer_stream, &mut peer_buf[..len]).await.unwrap();
-        let block = if let Payload::SyncBlock(block) = bincode::deserialize(&payload).unwrap() {
-            block
-        } else {
-            unreachable!();
-        };
-        let block = snarkvm_objects::Block::deserialize(&block).unwrap();
-
-        assert_eq!(block, block_struct_1);
-    }
-}
